Fix cyclic gx dep
